requirements.txt
# DID++ Backend Dependencies

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
# SQLite is built into Python, no extra dependency needed

# Encryption
cryptography==41.0.7

# ML/AI
mediapipe==0.10.9
opencv-python==4.8.1.78
numpy==1.26.2
librosa==0.10.1
easyocr==1.7.1

# Blockchain
web3==6.13.0
eth-account==0.10.0

# Data validation
pydantic==2.5.2
